+++ shellerate/encoder/xority.py
@@ -5,6 +5,7 @@ import logging
 
 from shellerate.encoder.encoder import Encoder
 from binascii import unhexlify, hexlify;
+from shellerate.utils import strings;
 
 # First version with clear text decoder stub: https://www.virustotal.com/#/file/7b25b33a1527d2285ebdefd327bc72b6d932c140489e8bfb7424bef115aa2ecd/detection
 
@@ -17,7 +18,7 @@ class Xority(Encoder):
   # calc_xor_key("d2b00bcd") -> 62bbc61f
   def calc_xor_key(self, val):
     key = ""
-    a=self.split(val)
+    a=strings.split(val)
     b=[]
 
     b.append(self.xor_str(a[0], secrets.token_hex(1)))
@@ -28,10 +29,10 @@ class Xority(Encoder):
     return ''.join(b)
 
   def encode(self, output_format="c"):
-    padded_shellcode = self.pad(self.shellcode())
+    padded_shellcode = strings.pad(self.shellcode())
     output = ""
     for c in textwrap.wrap(padded_shellcode, 16):
-      val=self.from_char_to_hexcode(c)
+      val=strings.from_char_to_hexcode(c)
       key=self.calc_xor_key(val)
       encoded=self.xor_str(val, key)
 
@@ -47,7 +48,7 @@ class Xority(Encoder):
     mark=secrets.token_hex(4)
     output+=mark
     output+=mark
-    a=self.split(output)
+    a=strings.split(output)
 
     if output_format == "c":
       return ''.join('\\x'+x.zfill(2) for x in a)
@@ -63,7 +64,7 @@ class Xority(Encoder):
     if output_format == "raw":
       return stub_raw + self.encode("raw")
     
-    a=self.split(stub_raw)
+    a=strings.split(stub_raw)
     if output_format == "c":
       stub=''.join('\\x'+x.zfill(2) for x in a)
     if output_format == "asm":