+++ shellerate/utils/asm_x86.py
@@ -1,0 +1,4 @@
+def get_esp_address_in_eax():
+  # PUSH ESP
+  # POP EAX
+  return "\\x54\\x58"